setup of the appdbContext

+++ Models/PCs.cs
@@ -4,7 +4,7 @@ public class PCs
 {
     public int Id { get; set; }
     public string Name { get; set; }
-    public decimal Weight { get; set; }
+    public float Weight { get; set; }
     public int Warranty { get; set; }
     public DateTime CreatedAt { get; set; }
     public int Stock { get; set; }
